+++ JLPTVoca/JLPTVoca/Model/DataLoader.swift
