+++ JLPTVoca/JLPTVoca/Model/RubyText.swift
