+++ JLPTVoca/JLPTVoca/View/RubyTextView.swift
